fetch from s3, fallback to overpass
